move `DispatcherProvider` from `core:test` to `core:base`
* move `DispatcherProvider` and `PlatformDispatcherProvider` to `core:base`
* move `dispatcherModule` to `core:base`
* update imports and gradle dependencies across modules
