+++ app/build.gradle.kts
@@ -106,7 +106,6 @@ dependencies {
         implementation(core.sync)
         implementation(core.base)
         implementation(core.uimarket)
-        implementation(core.test)
         implementation(domain.market)
         implementation(data.marketLocal)
         implementation(data.marketRemote)
