+++ domain/market/build.gradle.kts
@@ -58,7 +58,6 @@ dependencies {
     testImplementation(kotlin("test"))
 
     projects.apply {
-        api(core.test)
         api(core.network.ktor)
         testImplementation(testFixtures(core.test))
     }
