+++ feature/marketlist/src/main/java/ir/composenews/marketlist/MarketListViewModel.kt
@@ -7,7 +7,7 @@ import androidx.paging.cachedIn
 import androidx.paging.map
 import ir.composenews.base.BaseViewModel
 import ir.composenews.base.LoadableData
-import ir.composenews.core_test.dispatcher.DispatcherProvider
+import ir.composenews.base.dispatcher.DispatcherProvider
 import ir.composenews.domain.use_case.GetFavoriteMarketListUseCase
 import ir.composenews.domain.use_case.GetPagedMarketListUseCase
 import ir.composenews.domain.use_case.ToggleFavoriteMarketListUseCase
